requirements.txt
Flask
requests
huggingface_hub
python-dotenv
